Set image sizes on theme activation, not theme setup
The after_setup_theme hook fires on every pageload, but we don't need to be writing all these options to the database that often. after_switch_theme only fires when the theme is activated. There are some other options we could move here as well but we'll start with the image sizes.